add features

+++ src/Controller/ProductController.php
@@ -9,7 +9,7 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/product')]
 final class ProductController extends AbstractController
@@ -17,15 +17,27 @@ final class ProductController extends AbstractController
     #[Route(name: 'app_product_index', methods: ['GET'])]
     public function index(ProductRepository $productRepository): Response
     {
+        $user = $this->getUser();
+        if (in_array('ROLE_ADMIN', $user->getRoles(), true)) {
+            $products = $productRepository->findAll();
+        } else {
+            $products = $productRepository->findBy(['createdBy' => $user]);
+        }
+
         return $this->render('product/index.html.twig', [
-            'products' => $productRepository->findAll(),
+            'products' => $products,
         ]);
     }
 
+    #[IsGranted('ROLE_STAFF')]
     #[Route('/new', name: 'app_product_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $product = new Product();
+        $product->setCreatedBy($this->getUser());
+        $product->setCreatedAt(new \DateTimeImmutable());
+        $product->setUpdatedAt(new \DateTimeImmutable());
+
         $form = $this->createForm(ProductType::class, $product);
         $form->handleRequest($request);
 
@@ -45,6 +57,11 @@ final class ProductController extends AbstractController
     #[Route('/{id}', name: 'app_product_show', methods: ['GET'])]
     public function show(Product $product): Response
     {
+        $user = $this->getUser();
+        if (!in_array('ROLE_ADMIN', $user->getRoles(), true) && $product->getCreatedBy() !== $user) {
+            throw $this->createAccessDeniedException('You can only view your own products.');
+        }
+
         return $this->render('product/show.html.twig', [
             'product' => $product,
         ]);
@@ -53,10 +70,16 @@ final class ProductController extends AbstractController
     #[Route('/{id}/edit', name: 'app_product_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Product $product, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->getUser();
+        if (!in_array('ROLE_ADMIN', $user->getRoles(), true) && $product->getCreatedBy() !== $user) {
+            throw $this->createAccessDeniedException('You can only edit your own products.');
+        }
+
         $form = $this->createForm(ProductType::class, $product);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $product->setUpdatedAt(new \DateTimeImmutable());
             $entityManager->flush();
 
             return $this->redirectToRoute('app_product_index', [], Response::HTTP_SEE_OTHER);
@@ -71,6 +94,11 @@ final class ProductController extends AbstractController
     #[Route('/{id}', name: 'app_product_delete', methods: ['POST'])]
     public function delete(Request $request, Product $product, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->getUser();
+        if (!in_array('ROLE_ADMIN', $user->getRoles(), true) && $product->getCreatedBy() !== $user) {
+            throw $this->createAccessDeniedException('You can only delete your own products.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$product->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($product);
             $entityManager->flush();
